Try adding the rest of the dependencies
It is not clear to me why this would be required, but the error message is quite opaque to me and a similar change was necessary during the setup.py->pyproject.toml change for building with setup.py, so maybe it will work here. Further, I am monkey-see-monkey-doing compared to pympipool, which also includes its dependency suite in the `host` section, as this has been a template for me in the pyproject.toml transition. I'm still excluding the openjdk requirement, as this is (hopefully) only necessary for actually running the code.